further work implementing hue ring

+++ heavenLi_pyopengl/arnUtils/arnCircle.cpp
@@ -17,7 +17,7 @@ PyObject* drawHomeCircle_drawArn(PyObject *self, PyObject *args) {
    PyObject* py_list;
    PyObject* py_tuple;
    PyObject* py_float;
-   double *bulbColors;
+   float *bulbColors;
    float gx, gy, wx, wy, ao, w2h, R, G, B;
    int numBulbs;
    if (!PyArg_ParseTuple(args,
@@ -36,14 +36,14 @@ PyObject* drawHomeCircle_drawArn(PyObject *self, PyObject *args) {
    char circleSegments = 60;
 
    // Parse array of tuples containing RGB Colors of bulbs
-   bulbColors = new double[numBulbs*3];
+   bulbColors = new float[numBulbs*3];
 //#  pragma omp parallel for
    for (int i = 0; i < numBulbs; i++) {
       py_tuple = PyList_GetItem(py_list, i);
 
       for (int j = 0; j < 3; j++) {
          py_float = PyTuple_GetItem(py_tuple, j);
-         bulbColors[i*3+j] = double(PyFloat_AsDouble(py_float));
+         bulbColors[i*3+j] = float(PyFloat_AsDouble(py_float));
       }
    }
 
@@ -178,8 +178,8 @@ PyObject* drawIconCircle_drawArn(PyObject *self, PyObject *args) {
    PyObject* py_list;
    PyObject* py_tuple;
    PyObject* py_float;
-   double *bulbColors;
-   double detailColor[3];
+   float *bulbColors;
+   float detailColor[3];
    float gx, gy, scale, ao, w2h;
    int numBulbs, features;
    int vertIndex = 0;
@@ -201,21 +201,21 @@ PyObject* drawIconCircle_drawArn(PyObject *self, PyObject *args) {
    char circleSegments = 60;
 
    // Parse array of tuples containing RGB Colors of bulbs
-   bulbColors = new double[numBulbs*3];
+   bulbColors = new float[numBulbs*3];
 //#  pragma omp parallel for
    for (int i = 0; i < numBulbs; i++) {
       py_tuple = PyList_GetItem(py_list, i);
 
       for (int j = 0; j < 3; j++) {
          py_float = PyTuple_GetItem(py_tuple, j);
-         bulbColors[i*3+j] = double(PyFloat_AsDouble(py_float));
+         bulbColors[i*3+j] = float(PyFloat_AsDouble(py_float));
       }
    }
 
    // Parse RGB detail colors
-   detailColor[0] = PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 0));
-   detailColor[1] = PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 1));
-   detailColor[2] = PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 2));
+   detailColor[0] = float(PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 0)));
+   detailColor[1] = float(PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 1)));
+   detailColor[2] = float(PyFloat_AsDouble(PyTuple_GetItem(detailColorPyTup, 2)));
 
    if (iconCircleVertexBuffer == NULL     ||
        iconCircleColorBuffer  == NULL     ||
